+++ src/loader/loader.py
@@ -93,14 +93,33 @@ def get_omniglot(config, target_alphabets=[]):
 
     return background_data, evaluation_data
 
-def string_contains(string, set):
+def string_contains(string: str, set: list[str]):
     contains = False
     for entry in set:
         contains = contains or string.startswith(entry)
     
     return contains
 
+# Create loaders for each class in support data 
+# with batch size of shots.
+# Creates loader for queries which contain the rest of the data
+def k_shot_loaders(support_data, shots, query_batch_size=100):
+    all_indexs_to_remove = []
+    support_loaders = []
+    targets = torch.unique(support_data.targets)
+    
+    for target in targets:
+        subset_indexs = [j for j, x in enumerate(support_data.targets) if x == target][:shots]
+        all_indexs_to_remove.extend(subset_indexs)
+        subset_data = Subset(support_data, subset_indexs)
+        support_loaders.append(torch.utils.data.DataLoader(subset_data, batch_size=shots, shuffle=True, num_workers=1))
 
+    indexs_to_keep = [i for i in range(len(support_data.data)) if i not in all_indexs_to_remove]
+    query_data = Subset(support_data, indexs_to_keep)
+    
+    query_loader = torch.utils.data.DataLoader(query_data, batch_size=query_batch_size, shuffle=True, num_workers=1)
+
+    return support_loaders, query_loader
 
 if __name__ == '__main__':
     train, test = get_omniglot(0, ["Arcadian", "Armenian"])
